Move intervention objects to root module (#4859)
* Move intervention to root

* usage
