+++ python/packages/autogen-core/src/autogen_core/base/intervention.py
@@ -1,45 +1,22 @@
-from typing import Any, Awaitable, Callable, Protocol, final
+from typing_extensions import deprecated
 
-from .._agent_id import AgentId
+from .._intervention import DefaultInterventionHandler as DefaultInterventionHandlerAlias
+from .._intervention import DropMessage as DropMessageAlias
+from .._intervention import InterventionHandler as InterventionHandlerAliass
 
 __all__ = [
     "DropMessage",
-    "InterventionFunction",
     "InterventionHandler",
     "DefaultInterventionHandler",
 ]
 
-
-@final
-class DropMessage: ...
-
-
-InterventionFunction = Callable[[Any], Any | Awaitable[type[DropMessage]]]
+# Final so can't inherit and deprecate
+DropMessage = DropMessageAlias
 
 
-class InterventionHandler(Protocol):
-    """An intervention handler is a class that can be used to modify, log or drop messages that are being processed by the :class:`autogen_core.base.AgentRuntime`.
-
-    Note: Returning None from any of the intervention handler methods will result in a warning being issued and treated as "no change". If you intend to drop a message, you should return :class:`DropMessage` explicitly.
-    """
-
-    async def on_send(self, message: Any, *, sender: AgentId | None, recipient: AgentId) -> Any | type[DropMessage]: ...
-    async def on_publish(self, message: Any, *, sender: AgentId | None) -> Any | type[DropMessage]: ...
-    async def on_response(
-        self, message: Any, *, sender: AgentId, recipient: AgentId | None
-    ) -> Any | type[DropMessage]: ...
+@deprecated("Moved to autogen_core.InterventionHandler. Will remove this in 0.4.0.", stacklevel=2)
+class InterventionHandler(InterventionHandlerAliass): ...
 
 
-class DefaultInterventionHandler(InterventionHandler):
-    """Simple class that provides a default implementation for all intervention
-    handler methods, that simply returns the message unchanged. Allows for easy
-    subclassing to override only the desired methods."""
-
-    async def on_send(self, message: Any, *, sender: AgentId | None, recipient: AgentId) -> Any | type[DropMessage]:
-        return message
-
-    async def on_publish(self, message: Any, *, sender: AgentId | None) -> Any | type[DropMessage]:
-        return message
-
-    async def on_response(self, message: Any, *, sender: AgentId, recipient: AgentId | None) -> Any | type[DropMessage]:
-        return message
+@deprecated("Moved to autogen_core.DefaultInterventionHandler. Will remove this in 0.4.0.", stacklevel=2)
+class DefaultInterventionHandler(DefaultInterventionHandlerAlias): ...
